[refactor] 패키지 이동 및 code reformat
전체 패키지 domain과 view로 변경
code reformat
optimize imports
