+++ src/main/java/racingcar/domain/car/strategy/RandomNumMovingStrategy.java
@@ -1,4 +1,4 @@
-package racingcar.car.strategy;
+package racingcar.domain.car.strategy;
 
 import java.util.Random;
